+++ src/main/java/com/school/exam/web/account/UserAdminController.java
@@ -9,6 +9,8 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import com.school.exam.entity.TeExamPaperResultVO;
+import com.school.exam.repository.ExamPaperResultDao;
 import org.apache.log4j.spi.LoggerFactory;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,8 +39,10 @@ public class UserAdminController {
 	Logger logger = org.slf4j.LoggerFactory.getLogger(UserAdminController.class);
 	@Autowired
 	private AccountService accountService;
-	@Autowired
+    @Autowired
 	private SSClassService classService;
+    @Autowired
+    private ExamPaperResultDao resultDao;
 
 	@RequestMapping(method = RequestMethod.GET)
 	public String list(Model model) {
@@ -52,7 +56,10 @@ public class UserAdminController {
 	public String updateForm(@PathVariable("id") Long id, Model model) {
 		User user = accountService.getUser(id);
 		List<SSClassVO> sclass = classService.getAllSSClass();
-		model.addAttribute("user", accountService.getUser(id));
+        SSClassVO cvo = new SSClassVO();
+        cvo.setId(user.getSsClass().getId());
+        user.setSsClass(cvo);
+		model.addAttribute("user", user);
 		model.addAttribute("classLists", sclass);
 		model.addAttribute("pathUrl","更新用户");
 		return "account/adminUserForm";
@@ -60,6 +67,11 @@ public class UserAdminController {
 
 	@RequestMapping(value = "update", method = RequestMethod.POST)
 	public String update(@Valid @ModelAttribute("user") User user, RedirectAttributes redirectAttributes) {
+        SSClassVO classVO = new SSClassVO();
+        logger.warn("===id==={}", classVO.getId());
+        logger.warn("====name=={}", classVO.getClassName());
+        classVO = classService.getClass(user.getSsClass().getId());
+        user.setSsClass(classVO);
 		accountService.updateUser(user);
 		redirectAttributes.addFlashAttribute("message", "更新用户" + user.getLoginName() + "成功");
 		return "redirect:/admin/user";
@@ -77,9 +89,14 @@ public class UserAdminController {
 	@RequestMapping(value = "delete/{id}")
 	public String delete(@PathVariable("id") Long id, RedirectAttributes redirectAttributes) {
 		User user = accountService.getUser(id);
-		accountService.deleteUser(id);
-		redirectAttributes.addFlashAttribute("message", "删除用户" + user.getLoginName() + "成功");
-		return "redirect:/admin/user";
+       List<TeExamPaperResultVO> resultList = resultDao.findByPersonId(user.getId());
+        if(null!=resultList&&resultList.size()>0){
+            redirectAttributes.addFlashAttribute("message", "当前用户" + user.getLoginName() + "已经参加考试，不能进行删除操作");
+        }else{
+            accountService.deleteUser(id);
+            redirectAttributes.addFlashAttribute("message", "删除用户" + user.getLoginName() + "成功");
+        }
+        return "redirect:/admin/user";
 	}
 
 	/**
